requirements.txt
langgraph>=0.2.0
langchain-openai>=0.3.0
langchain-core>=0.3.0
httpx>=0.27.0
beautifulsoup4>=4.12.0
pydantic>=2.0.0
python-dotenv>=1.0.0
pyyaml>=6.0.0
